hotfix: 토큰 재발급 로직 수정
